Added denoising functionality

+++ src/materials.h
@@ -25,11 +25,11 @@ struct MicrofacetSampleArgs{
     vec3 normal_vector;
     vec3 incident_vector;
     vec3 intersection_point;
-    double cosine_factor;
-    double eta;
-    double u;
-    double v;
-    double alpha;
+    float cosine_factor;
+    float eta;
+    float u;
+    float v;
+    float alpha;
     Object** objects;
     int number_of_objects;
     bool outside;
@@ -38,19 +38,19 @@ struct MicrofacetSampleArgs{
 
 struct MicrofacetData{
     bool outside;
-    double alpha;
-    double eta;
+    float alpha;
+    float eta;
     vec3 normal_into_interface;
     vec3 half_vector;
-    double F_r;
+    float F_r;
 };
 
 
 struct MaterialData{
     ValueMap3D* albedo_map = nullptr;
-    double refractive_index = 1;
-    double extinction_coefficient = 0;
-    double attenuation_coefficient = 0;
+    float refractive_index = 1;
+    float extinction_coefficient = 0;
+    float attenuation_coefficient = 0;
     vec3 absorption_albedo = colors::WHITE;
     ValueMap3D* emission_color_map = nullptr;
     ValueMap1D* light_intensity_map = nullptr;
@@ -64,13 +64,13 @@ struct MaterialData{
 class Material{
     public:
         ValueMap3D* albedo_map;
-        double refractive_index;
-        double attenuation_coefficient;
+        float refractive_index;
+        float attenuation_coefficient;
         vec3 absorption_albedo;
         ValueMap3D* emission_color_map;
         ValueMap1D* light_intensity_map;
         bool is_dielectric;
-        double extinction_coefficient;
+        float extinction_coefficient;
         bool is_light_source;
         ValueMap1D* roughness_map;
         ValueMap1D* percentage_diffuse_map;
@@ -123,19 +123,19 @@ class Material{
         delete percentage_diffuse_map;
     }
 
-    virtual vec3 eval(const Hit& hit, const double u, const double v){
+    virtual vec3 eval(const Hit& hit, const float u, const float v){
         throw VirtualMethodNotAllowedException("This is a pure virtual method and should not be called.");
         vec3 vec;
         return vec;
     }
 
-    virtual BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const double u, const double v){
+    virtual BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const float u, const float v){
         throw VirtualMethodNotAllowedException("This is a pure virtual method and should not be called.");
         BrdfData data;
         return data;
     }
 
-    vec3 get_light_emittance(const double u, const double v){
+    vec3 get_light_emittance(const float u, const float v){
         return emission_color_map -> get(u, v) * light_intensity_map -> get(u, v);
     }
 };
@@ -145,11 +145,11 @@ class DiffuseMaterial : public Material{
     public:
         using Material::Material;
 
-    vec3 eval(const Hit& hit, const double u, const double v) override{
+    vec3 eval(const Hit& hit, const float u, const float v) override{
         return albedo_map -> get(u, v) / M_PI;
     }
 
-    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const double u, const double v) override{
+    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const float u, const float v) override{
         vec3 adjusted_normal = dot_vectors(hit.incoming_vector, hit.normal_vector) < 0 ? hit.normal_vector : -hit.normal_vector;
         vec3 outgoing_vector = sample_cosine_hemisphere(adjusted_normal);
         vec3 brdf_multiplier = albedo_map -> get(u, v);
@@ -166,11 +166,11 @@ class ReflectiveMaterial : public Material{
     public:
         using Material::Material;
 
-    vec3 eval(const Hit& hit, const double u, const double v) override{
+    vec3 eval(const Hit& hit, const float u, const float v) override{
         return colors::BLACK;
     }
 
-    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const double u, const double v) override{
+    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const float u, const float v) override{
         vec3 adjusted_normal = dot_vectors(hit.incoming_vector, hit.normal_vector) < 0 ? hit.normal_vector : -hit.normal_vector;
         vec3 outgoing_vector = reflect_vector(hit.incoming_vector, adjusted_normal);
         BrdfData data;
@@ -186,18 +186,18 @@ class TransparentMaterial : public Material{
     public:
         using Material::Material;
 
-    vec3 eval(const Hit& hit, const double u, const double v) override{
+    vec3 eval(const Hit& hit, const float u, const float v) override{
         return colors::BLACK;
     }
 
-    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const double u, const double v) override{
-        double incoming_dot_normal = dot_vectors(hit.incoming_vector, hit.normal_vector);
+    BrdfData sample(const Hit& hit, Object** scene_objects, const int number_of_objects, const float u, const float v) override{
+        float incoming_dot_normal = dot_vectors(hit.incoming_vector, hit.normal_vector);
         vec3 normal_into_interface;
         bool inside = incoming_dot_normal > 0.0;
-        double n1;
-        double k1;
-        double n2;
-        double k2;
+        float n1;
+        float k1;
+        float n2;
+        float k2;
         if (!inside){
             normal_into_interface = -hit.normal_vector;
             n1 = constants::air_refractive_index;
@@ -215,13 +215,13 @@ class TransparentMaterial : public Material{
 
         vec3 transmitted_vector = refract_vector(hit.incoming_vector, normal_into_interface, n1 / n2);
 
-        double F_r = 1;
+        float F_r = 1;
         if (transmitted_vector.length_squared() != 0){
-            double cos_incident = dot_vectors(hit.incoming_vector, normal_into_interface);
+            float cos_incident = dot_vectors(hit.incoming_vector, normal_into_interface);
             F_r = fresnel_multiplier(cos_incident, n1, k1, n2, k2, is_dielectric);
         }
 
-        double random_num = random_uniform(0, 1);
+        float random_num = random_uniform(0, 1);
         bool is_reflected = random_num <= F_r;
         
         BrdfData data;
@@ -237,7 +237,7 @@ class TransparentMaterial : public Material{
             transmission_ray.starting_position = hit.intersection_point;
             Hit transmission_hit = find_closest_hit(transmission_ray, scene_objects, number_of_objects);
             vec3 attenuation_color;
-            double distance = transmission_hit.distance;
+            float distance = transmission_hit.distance;
             if (distance > 0 && !inside){
                 vec3 log_attenuation = absorption_albedo * attenuation_coefficient * (-distance);
                 attenuation_color = exp_vector(log_attenuation);
@@ -259,29 +259,29 @@ class MicrofacetMaterial : public Material{
     public:
         using Material::Material;
 
-    double chi(const double x){
+    float chi(const float x){
         return x > 0 ? 1 : 0;
     }
 
-    double G1(const vec3& half_vector, const vec3& normal_vector, const vec3& v, const double alpha){
-        double cos_theta = dot_vectors(half_vector, v);
-        double tan_theta = sqrt((1.0 - cos_theta * cos_theta) / (cos_theta * cos_theta));
-        double a = 1.0 / (alpha * tan_theta);
+    float G1(const vec3& half_vector, const vec3& normal_vector, const vec3& v, const float alpha){
+        float cos_theta = dot_vectors(half_vector, v);
+        float tan_theta = sqrt((1.0 - cos_theta * cos_theta) / (cos_theta * cos_theta));
+        float a = 1.0 / (alpha * tan_theta);
         
         return chi(cos_theta / dot_vectors(v, normal_vector)) * a < 1.6 ? (3.535 * a  + 2.181 * a * a) / (1.0 + 2.276 * a + 2.577 * a * a) : 1.0;
     }
 
-    double G(const vec3& half_vector, const vec3& normal_vector, const vec3& incident_vector, const vec3& outgoing_vector, const double alpha){
+    float G(const vec3& half_vector, const vec3& normal_vector, const vec3& incident_vector, const vec3& outgoing_vector, const float alpha){
         return G1(half_vector, normal_vector, incident_vector, alpha) * G1(half_vector, normal_vector, outgoing_vector, alpha);
     }
 
-    MicrofacetData prepare_microfacet_data(const Hit& hit, const double u, const double v){
-        double n1;
-        double k1;
-        double n2;
-        double k2;
+    MicrofacetData prepare_microfacet_data(const Hit& hit, const float u, const float v){
+        float n1;
+        float k1;
+        float n2;
+        float k2;
         vec3 normal_into_interface;
-        double incoming_dot_normal = dot_vectors(hit.incoming_vector, hit.normal_vector);
+        float incoming_dot_normal = dot_vectors(hit.incoming_vector, hit.normal_vector);
         bool outside = incoming_dot_normal <= 0.0;
         if (outside){
             normal_into_interface = -hit.normal_vector;
@@ -298,13 +298,13 @@ class MicrofacetMaterial : public Material{
             k2 = 0;
         }
 
-        double alpha = std::max(roughness_map -> get(u, v), constants::EPSILON);
+        float alpha = std::max(roughness_map -> get(u, v), constants::EPSILON);
 
         vec3 sampled_half_vector = specular_sample(-normal_into_interface, alpha);
                 
-        double i_dot_h = dot_vectors(hit.incoming_vector, sampled_half_vector);
+        float i_dot_h = dot_vectors(hit.incoming_vector, sampled_half_vector);
 
-        double F_r = fresnel_multiplier(-i_dot_h, n1, k1, n2, k2, is_dielectric);
+        float F_r = fresnel_multiplier(-i_dot_h, n1, k1, n2, k2, is_dielectric);
 
         MicrofacetData data;
         data.outside = outside;
@@ -316,15 +316,15 @@ class MicrofacetMaterial : public Material{
         return data;
     }
 
-    vec3 eval(const Hit& hit, const double u, const double v) override{
+    vec3 eval(const Hit& hit, const float u, const float v) override{
         MicrofacetData data = prepare_microfacet_data(hit, u, v);
         
-        double random_num = random_uniform(0, 1);
+        float random_num = random_uniform(0, 1);
         bool reflect_specular = random_num < data.F_r;
         if (reflect_specular){
             return colors::BLACK;
         }
-        double transmit = random_uniform(0, 1) > percentage_diffuse_map -> get(u, v);
+        float transmit = random_uniform(0, 1) > percentage_diffuse_map -> get(u, v);
 
         if (transmit){
             return colors::BLACK;
@@ -333,15 +333,15 @@ class MicrofacetMaterial : public Material{
         return albedo_map -> get(u, v) / M_PI;
     }
 
-    vec3 specular_sample(const vec3& normal_vector, const double alpha){
-        double r1 = random_uniform(0, 1);
-        double r2 = random_uniform(0, 1);
-        double phi = 2 * M_PI * r2;
-        double tan_theta2 = - alpha * alpha * std::log(1 - r1);
-        double cos_theta2 = 1.0 / (1.0 + tan_theta2);
+    vec3 specular_sample(const vec3& normal_vector, const float alpha){
+        float r1 = random_uniform(0, 1);
+        float r2 = random_uniform(0, 1);
+        float phi = 2 * M_PI * r2;
+        float tan_theta2 = - alpha * alpha * std::log(1 - r1);
+        float cos_theta2 = 1.0 / (1.0 + tan_theta2);
         
-        double cos_theta = sqrt(cos_theta2);
-        double sin_theta = sqrt(1 - cos_theta2);
+        float cos_theta = sqrt(cos_theta2);
+        float sin_theta = sqrt(1 - cos_theta2);
 
         vec3 x_hat;
         vec3 y_hat;
@@ -372,7 +372,7 @@ class MicrofacetMaterial : public Material{
         transmission_ray.starting_position = args.intersection_point;
         Hit transmission_hit = find_closest_hit(transmission_ray, args.objects, args.number_of_objects);
         vec3 attenuation_color;
-        double distance = transmission_hit.distance;
+        float distance = transmission_hit.distance;
         if (distance > 0 && args.outside){
             vec3 log_attenuation = absorption_albedo * attenuation_coefficient * (-distance);
             attenuation_color = exp_vector(log_attenuation);
@@ -402,16 +402,16 @@ class MicrofacetMaterial : public Material{
     }
     
     
-    BrdfData sample(const Hit& hit, Object** objects, const int number_of_objects, const double u, const double v) override{
+    BrdfData sample(const Hit& hit, Object** objects, const int number_of_objects, const float u, const float v) override{
         MicrofacetData data = prepare_microfacet_data(hit, u, v);
                 
-        double i_dot_h = dot_vectors(hit.incoming_vector, data.half_vector);
-        double i_dot_n = dot_vectors(hit.incoming_vector, data.normal_into_interface);
-        double n_dot_h = dot_vectors(data.half_vector, data.normal_into_interface);
+        float i_dot_h = dot_vectors(hit.incoming_vector, data.half_vector);
+        float i_dot_n = dot_vectors(hit.incoming_vector, data.normal_into_interface);
+        float n_dot_h = dot_vectors(data.half_vector, data.normal_into_interface);
 
-        double cosine_factor = std::abs(i_dot_h / (i_dot_n * n_dot_h));
+        float cosine_factor = std::abs(i_dot_h / (i_dot_n * n_dot_h));
 
-        double random_num = random_uniform(0, 1);
+        float random_num = random_uniform(0, 1);
         bool reflect_specular = random_num < data.F_r;
         
         MicrofacetSampleArgs args;
@@ -431,7 +431,7 @@ class MicrofacetMaterial : public Material{
             args.objects = objects;
             args.number_of_objects = number_of_objects;
             args.outside = data.outside;
-            double random_num2 = random_uniform(0, 1);
+            float random_num2 = random_uniform(0, 1);
             bool diffuse = random_num2 < percentage_diffuse_map -> get(u, v);
             return diffuse ? sample_diffuse(args) : sample_transmission(args);
         }
